implement controller, fix path to not include root, use bytes for bus and registers

+++ flip/components/bus.py
@@ -1,5 +1,6 @@
 from typing import Iterable, Optional, override
 
+from flip.bytes import Byte
 from flip.components import component
 
 
@@ -15,7 +16,7 @@ class Bus(component.Component):
         children: Optional[Iterable[component.Component]] = None,
     ) -> None:
         super().__init__(name=name, parent=parent, children=children)
-        self.__value: Optional[int] = None
+        self.__value: Optional[Byte] = None
         self.__setter: Optional[str] = None
 
     @override
@@ -23,14 +24,14 @@ class Bus(component.Component):
         return f"Bus(name={self.name}, value={self.value}, setter={self.setter})"
 
     @property
-    def value(self) -> Optional[int]:
+    def value(self) -> Optional[Byte]:
         return self.__value
 
     @property
     def setter(self) -> Optional[str]:
         return self.__setter
 
-    def set(self, value: int, setter: str | component.Component) -> None:
+    def set(self, value: Byte, setter: str | component.Component) -> None:
         if isinstance(setter, component.Component):
             setter = setter.path
         if self.__setter is not None and setter != self.__setter:
